complete rest of form refactor/decouple validation
