+++ src/features/testResults/Form.jsx
@@ -1,14 +1,18 @@
 import { FormProvider, useForm } from 'react-hook-form';
 import AgeSelect from '../userDetails/AgeSelect';
 import GenderSelect from '../userDetails/GenderSelect';
-import { DevTool } from '@hookform/devtools';
 
-import React, { useContext } from 'react';
-import ExerciseForm from './ExerciseFormRefactor';
-import { getScores } from '../../api/supabase';
+import React from 'react';
+import ExerciseForm from './ExerciseForm';
+import { fetchExerciseScores } from '../../api/supabase';
 import Score from '../../ui/Score';
 import { useScoreContext } from '../../contexts/ScoreContext';
 import FormButtons from '../../ui/FormButtons';
+import {
+  cardioValidationRules,
+  coreValidationRules,
+  upperBodyValidationRules,
+} from './validationRules';
 
 const InputStyle =
   'w-full rounded-full p-5 text-center font-semibold text-stone-700 shadow-lg';
@@ -18,20 +22,20 @@ function Form() {
   const { register, watch } = methods;
 
   const {
-    setUpperScore,
+    setUpperBodyScore,
     setCoreScore,
     setCardioScore,
     setTotalScore,
     setIsMinimumNotMet,
   } = useScoreContext();
 
-  async function onSubmit(data) {
+  async function handleFormSubmit(data) {
     setIsMinimumNotMet(false);
-    const { upper, core, cardio } = await getScores(data);
+    const { upper, core, cardio } = await fetchExerciseScores(data);
     if (upper === 0 || core === 0 || cardio === 0) {
       setIsMinimumNotMet(true);
     }
-    setUpperScore(upper);
+    setUpperBodyScore(upper);
     setCoreScore(core);
     setCardioScore(cardio);
     setTotalScore(upper + core + cardio);
@@ -41,7 +45,7 @@ function Form() {
     <FormProvider {...methods}>
       <form
         className="m-auto mb-3 flex max-w-2xl flex-col gap-24 text-2xl  uppercase tracking-widest text-stone-200 sm:max-w-3xl"
-        onSubmit={methods.handleSubmit(onSubmit)}
+        onSubmit={methods.handleSubmit(handleFormSubmit)}
         noValidate
       >
         <GenderSelect />
@@ -63,15 +67,10 @@ function Form() {
             <ExerciseForm.InputHeader>Upper Body Reps</ExerciseForm.InputHeader>
             <ExerciseForm.Input>
               <input
-                {...register('upperBodyResults', {
-                  required: 'Rep amount is required',
-                  min: { value: 0, message: 'Reps must be greater than 0' },
-                  max: { value: 125, message: 'Maximum amount exceeded' },
-                  pattern: {
-                    value: /^\d*$/,
-                    message: 'Must be a whole number',
-                  },
-                })}
+                {...register(
+                  'upperBodyResults',
+                  upperBodyValidationRules(watch('upperBodyExercise', '')),
+                )}
                 className={InputStyle}
                 type="number"
                 placeholder="Reps"
@@ -92,56 +91,60 @@ function Form() {
             </ExerciseForm.Dropdown>
           </div>
 
-          {/* This wrapper makes it so that the input only appears when a dropdown option is selected */}
           <ExerciseForm.InputVisibilityWrapper>
-            <ExerciseForm.InputHeader />
+            <ExerciseForm.InputHeader>{`Core ${
+              watch('coreExercise', '') === 'plank' ? 'Plank Time' : 'Reps'
+            }`}</ExerciseForm.InputHeader>
             <ExerciseForm.Input>
               <input
                 {...register(
                   'coreResults',
-                  watch('coreExercise', '') === 'plank'
-                    ? {
-                        required: 'Plank time is required',
-                        pattern: {
-                          value: /^(([0]?[0-5][0-9]|[0-9]):([0-5][0-9]))$/,
-                          message: 'Invalid time format. (mm:ss)',
-                        },
-                        min: {
-                          value: 0,
-                          message: 'Time must be greater than 0',
-                        },
-                      }
-                    : {
-                        required: 'Rep amount is required',
-                        min: {
-                          value: 0,
-                          message: 'Reps must be greater than 0',
-                        },
-                        max: { value: 125, message: 'Maximum amount exceeded' },
-                        pattern: {
-                          value: /^\d*$/,
-                          message: 'Must be a whole number',
-                        },
-                      },
+                  coreValidationRules(watch('coreExercise'), ''),
                 )}
                 className={InputStyle}
-                type="number"
-                placeholder="Reps"
+                type={watch('coreExercise') === 'plank' ? 'text' : 'number'}
+                placeholder={
+                  watch('coreExercise') === 'plank' ? 'mm:ss' : 'Reps'
+                }
               />
             </ExerciseForm.Input>
           </ExerciseForm.InputVisibilityWrapper>
         </ExerciseForm>
 
-        {/* <ExerciseForm type="core">
-          <option value="situps">Situps</option>
-          <option value="crunches">Cross Legged Reverse Crunch</option>
-          <option value="plank">Forearm Plank</option>
-        </ExerciseForm>
-
         <ExerciseForm type="cardio">
-          <option value="mile">1.5 Mile Run</option>
-          <option value="shuttles">20 Meter HAMR Shuttle</option>
-        </ExerciseForm> */}
+          <div>
+            <ExerciseForm.DropdownHeader>
+              Cardio Exercise
+            </ExerciseForm.DropdownHeader>
+            <ExerciseForm.Dropdown>
+              <option value="mile">1.5 Mile Run</option>
+              <option value="shuttles">20 Meter HAMR Shuttle</option>
+            </ExerciseForm.Dropdown>
+          </div>
+
+          <ExerciseForm.InputVisibilityWrapper>
+            <ExerciseForm.InputHeader>
+              {watch('cardioExercise') === 'shuttles'
+                ? 'HAMR Reps'
+                : 'Run Time'}
+            </ExerciseForm.InputHeader>
+            <ExerciseForm.Input>
+              <input
+                {...register(
+                  'cardioResults',
+                  cardioValidationRules(watch('cardioExercise')),
+                )}
+                className={InputStyle}
+                type={
+                  watch('cardioExercise') === 'shuttles' ? 'number' : 'text'
+                }
+                placeholder={
+                  watch('cardioExercise') === 'shuttles' ? 'Reps' : 'mm:ss'
+                }
+              />
+            </ExerciseForm.Input>
+          </ExerciseForm.InputVisibilityWrapper>
+        </ExerciseForm>
 
         <FormButtons />
       </form>
